requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
defusedxml==0.7.1
django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser==2.3.3
ecdsa==0.19.1
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
markupsafe==3.0.2
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
prompt-toolkit==3.0.51
propcache==0.3.2
pycparser==2.22
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sendgrid==6.12.4
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
twilio==9.6.5
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==3.0.1
django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser==2.3.3
dotenv==0.9.9
ecdsa==0.19.1
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
markupsafe==3.0.2
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
prompt-toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sendgrid==6.12.4
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
twilio==9.6.5
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.1
